Add files via upload
A farmer has 150 acres of land on which he wants to grow cotton and potatoes. Government restrictions
prevent the farmer from devoting more than 60 acres of land to cotton, but he can use as much acreage
for potatoes as he wishes. It requires 30 inches per acre of water for irritation to grow cotton and 15
inches per acre to grow potatoes. The farmer has 3000 acre-inches of water available. If the farmer's
pro¯t per acre is $207 for cotton and $200 for potatoes, how many acres of cotton and how many acres
of potatoes should he grow in order to obtain the largest possible pro¯t from his land?